Small changes on player script
Made a condition on Input left and right if so the player velocity stays at maximum velocity when he reaches it.
Changed Vector2.DOWN to UP so the player never stays inside the ground, also made the Y in the snap 2 instead of 20, this makes sure the char can jump. And fixed some wrong codes, forgot to add 4 (number of maximum slides), and then MAX_FLOOR_ANGLE.
Changed gravity to work only when it's not detecting ground, this fixes player to not get sticked on the ground while trying to jump.